refactored repo for similar interface of both projects
