--- ase-coap-erko/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "ase-coap-erko",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev-server": "nodemon ./server",
-    "dev-client": "nodemon ./client",
-    "server": "node ./server",
-    "client": "node ./client"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "nodemon": "^2.0.6"
-  },
-  "dependencies": {
-    "coap": "^0.24.0"
-  }
-}
