+++ ase-coap-erko/server/index.js
@@ -1,6 +1,5 @@
 const coap = require('coap');
-
-const Util = require('../common');
+const Util = require('../../common');
 
 // Read key
 const KEY = process.env.KEY && process.env.KEY.length >= 16 ? process.env.KEY : null;
@@ -13,10 +12,13 @@ if (!KEY) {
 const server = coap.createServer({ type: 'udp4' });
 
 server.on('request', function(req, res) {
+  // Decrypt and parse what we got
   let payload = req.payload.toString('utf-8');
   let decrypted = Util.decrypt(payload, KEY);
   const data = (() => { try { return JSON.parse(decrypted); } catch (e) { return null; }})();
   console.log('Temp=', data ? data.temp : '?');
+
+  // Respond with encrypted data
   let response = JSON.stringify({ success: true, updated: { temp: data.temp } });
   let encryptedResponse = Util.encrypt(response, KEY);
   res.end(encryptedResponse);
