--- ase-mqtt-erko/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "ase-mqtt-erko",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev-client": "nodemon ./client",
-    "client": "node ./client"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "nodemon": "^2.0.6"
-  },
-  "dependencies": {
-    "mqtt": "^4.2.5"
-  }
-}
